+++ src/dawn_wire/client/Adapter.cpp
@@ -41,7 +41,7 @@ namespace dawn_wire::client {
         return mLimitsAndFeatures.HasFeature(feature);
     }
 
-    uint32_t Adapter::EnumerateFeatures(WGPUFeatureName* features) const {
+    size_t Adapter::EnumerateFeatures(WGPUFeatureName* features) const {
         return mLimitsAndFeatures.EnumerateFeatures(features);
     }
