+++ src/tests/unittests/wire/WireAdapterTests.cpp
@@ -193,7 +193,7 @@ namespace {
                         for (wgpu::FeatureName feature : fakeFeatures) {
                             *(features++) = static_cast<WGPUFeatureName>(feature);
                         }
-                        return static_cast<uint32_t>(fakeFeatures.size());
+                        return fakeFeatures.size();
                     })));
 
                 api.CallAdapterRequestDeviceCallback(apiAdapter, WGPURequestDeviceStatus_Success,
@@ -259,7 +259,7 @@ namespace {
                         for (wgpu::FeatureName feature : fakeFeatures) {
                             *(features++) = static_cast<WGPUFeatureName>(feature);
                         }
-                        return static_cast<uint32_t>(fakeFeatures.size());
+                        return fakeFeatures.size();
                     })));
 
                 // The device was actually created, but the wire didn't support its features.
